moved transaction_record_map implementations to transaction_records_by_id
