+++ zingolib/src/wallet/transactions.rs
@@ -1,55 +1,7 @@
 use std::collections::HashMap;
 
-use zcash_client_backend::PoolType;
-use zcash_primitives::transaction::TxId;
+use crate::wallet::{data::WitnessTrees, transaction_records_by_id::TransactionRecordsById};
 
-use crate::wallet::{
-    data::WitnessTrees, notes::NoteRecordIdentifier, traits::DomainWalletExt,
-    transaction_records_by_id::TransactionRecordsById,
-};
-
-use super::data::TransactionRecord;
-
-#[derive(Debug)]
-pub struct TransactionRecordMap(pub HashMap<TxId, TransactionRecord>);
-
-impl std::ops::Deref for TransactionRecordMap {
-    type Target = HashMap<TxId, TransactionRecord>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl std::ops::DerefMut for TransactionRecordMap {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-impl TransactionRecordMap {
-    pub fn get_received_note_from_identifier<D: DomainWalletExt>(
-        &self,
-        note_record_reference: NoteRecordIdentifier,
-    ) -> Option<
-        zcash_client_backend::wallet::ReceivedNote<
-            NoteRecordIdentifier,
-            <D as zcash_note_encryption::Domain>::Note,
-        >,
-    >
-    where
-        <D as zcash_note_encryption::Domain>::Note: PartialEq + Clone,
-        <D as zcash_note_encryption::Domain>::Recipient: super::traits::Recipient,
-    {
-        let transaction = self.get(&note_record_reference.txid);
-        transaction.and_then(|transaction_record| {
-            if note_record_reference.pool == PoolType::Shielded(D::protocol()) {
-                transaction_record.get_received_note::<D>(note_record_reference.index)
-            } else {
-                None
-            }
-        })
-    }
-}
 /// HashMap of all transactions in a wallet, keyed by txid.
 /// Note that the parent is expected to hold a RwLock, so we will assume that all accesses to
 /// this struct are threadsafe/locked properly.
@@ -61,7 +13,6 @@ pub struct TxMapAndMaybeTrees {
 pub mod get;
 pub mod read_write;
 pub mod recording;
-mod trait_inputsource;
 
 impl TxMapAndMaybeTrees {
     pub(crate) fn new_with_witness_trees() -> TxMapAndMaybeTrees {
