+++ zingolib/src/wallet/transactions/trait_inputsource.rs
@@ -1,139 +1,1 @@
-use orchard::note_encryption::OrchardDomain;
-use sapling_crypto::note_encryption::SaplingDomain;
-use zcash_client_backend::{
-    data_api::{InputSource, SpendableNotes},
-    wallet::ReceivedNote,
-    PoolType, ShieldedProtocol,
-};
-use zcash_primitives::{transaction::components::amount::NonNegativeAmount, zip32::AccountId};
 
-use crate::{
-    error::{ZingoLibError, ZingoLibResult},
-    wallet::notes::NoteRecordIdentifier,
-};
-
-use super::TransactionRecordMap;
-
-impl InputSource for TransactionRecordMap {
-    type Error = ZingoLibError;
-    type AccountId = zcash_primitives::zip32::AccountId;
-    type NoteRef = NoteRecordIdentifier;
-
-    fn get_spendable_note(
-        &self,
-        txid: &zcash_primitives::transaction::TxId,
-        protocol: zcash_client_backend::ShieldedProtocol,
-        index: u32,
-    ) -> Result<
-        Option<
-            zcash_client_backend::wallet::ReceivedNote<
-                Self::NoteRef,
-                zcash_client_backend::wallet::Note,
-            >,
-        >,
-        Self::Error,
-    > {
-        let note_record_reference: <Self as InputSource>::NoteRef = NoteRecordIdentifier {
-            txid: *txid,
-            pool: PoolType::Shielded(protocol),
-            index,
-        };
-        match protocol {
-            ShieldedProtocol::Sapling => Ok(self
-                .get_received_note_from_identifier::<SaplingDomain>(note_record_reference)
-                .map(|note| {
-                    note.map_note(|note_inner| {
-                        zcash_client_backend::wallet::Note::Sapling(note_inner)
-                    })
-                })),
-            ShieldedProtocol::Orchard => Ok(self
-                .get_received_note_from_identifier::<OrchardDomain>(note_record_reference)
-                .map(|note| {
-                    note.map_note(|note_inner| {
-                        zcash_client_backend::wallet::Note::Orchard(note_inner)
-                    })
-                })),
-        }
-    }
-
-    fn select_spendable_notes(
-        &self,
-        account: Self::AccountId,
-        target_value: zcash_primitives::transaction::components::amount::NonNegativeAmount,
-        sources: &[zcash_client_backend::ShieldedProtocol],
-        anchor_height: zcash_primitives::consensus::BlockHeight,
-        exclude: &[Self::NoteRef],
-    ) -> Result<SpendableNotes<NoteRecordIdentifier>, ZingoLibError> {
-        if account != AccountId::ZERO {
-            return Err(ZingoLibError::Error(
-                "we don't use non-zero accounts (yet?)".to_string(),
-            ));
-        }
-        let mut sapling_note_noteref_pairs: Vec<(sapling_crypto::Note, NoteRecordIdentifier)> =
-            Vec::new();
-        let mut orchard_note_noteref_pairs: Vec<(orchard::Note, NoteRecordIdentifier)> = Vec::new();
-        for transaction_record in self.values().filter(|transaction_record| {
-            transaction_record
-                .status
-                .is_confirmed_before_or_at(&anchor_height)
-        }) {
-            if sources.contains(&ShieldedProtocol::Sapling) {
-                sapling_note_noteref_pairs.extend(
-                    transaction_record
-                        .select_unspent_note_noteref_pairs_sapling()
-                        .into_iter()
-                        .filter(|note_ref_pair| !exclude.contains(&note_ref_pair.1)),
-                );
-            }
-            if sources.contains(&ShieldedProtocol::Orchard) {
-                orchard_note_noteref_pairs.extend(
-                    transaction_record
-                        .select_unspent_note_noteref_pairs_orchard()
-                        .into_iter()
-                        .filter(|note_ref_pair| !exclude.contains(&note_ref_pair.1)),
-                );
-            }
-        }
-        let mut sapling_notes =
-            Vec::<ReceivedNote<NoteRecordIdentifier, sapling_crypto::Note>>::new();
-        let mut orchard_notes = Vec::<ReceivedNote<NoteRecordIdentifier, orchard::Note>>::new();
-        if let Some(missing_value_after_sapling) = sapling_note_noteref_pairs.into_iter().rev(/*biggest first*/).try_fold(
-            Some(target_value),
-            |rolling_target, (note, noteref)| match rolling_target {
-                Some(targ) => {
-                    sapling_notes.push(
-                        self.get(&noteref.txid).map(|tr| tr.get_received_note::<SaplingDomain>(noteref.index)).flatten()
-                            .ok_or_else(|| ZingoLibError::Error("missing note".to_string()))?
-                    );
-                    Ok(targ
-                        - NonNegativeAmount::from_u64(note.value().inner())
-                            .map_err(|e| ZingoLibError::Error(e.to_string()))?)
-                }
-                None => Ok(None),
-            },
-        )? {
-            if let Some(missing_value_after_orchard) = orchard_note_noteref_pairs.into_iter().rev(/*biggest first*/).try_fold(
-            Some(missing_value_after_sapling),
-            |rolling_target, (note, noteref)| match rolling_target {
-                Some(targ) => {
-                    orchard_notes.push(
-                        self.get(&noteref.txid).map(|tr| tr.get_received_note::<OrchardDomain>(noteref.index)).flatten()
-                            .ok_or_else(|| ZingoLibError::Error("missing note".to_string()))?
-                    );
-                    Ok(targ
-                        - NonNegativeAmount::from_u64(note.value().inner())
-                            .map_err(|e| ZingoLibError::Error(e.to_string()))?)
-                }
-                None => Ok(None),
-            },
-        )? {
-                return ZingoLibResult::Err(ZingoLibError::Error(format!(
-                    "insufficient funds, short {}",
-                    missing_value_after_orchard.into_u64()
-                )));
-            };
-        };
-
-        Ok(SpendableNotes::new(sapling_notes, orchard_notes))
-    }
-}
